Move serializer + fix doc
